+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## [0.0.2 UNRELEASED]
 
-### Added
+### Changed
+- Renamed vars in `variables.tf` and where they are used #8
 
+### Added
 - Using Vault to generate username and password for postgres with kv #13
 - Added 03_run_tests.yml #3
 - Code to support successful execution of nomad mc job and tests when consul_acl_default_policy is deny #22
+- Added documentation and LICENSE #2
 
 ## [0.0.1]